move text wrap handling to separate file
